+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ A detailed list of all notable changes for every released version.
 ## [0.0.99](https://github.com/bloopgames/toodle/releases/tag/0.0.97)
 
 - Add low level api for post-process (fullscreen) shaders, see [post-process example](https://toodle.gg/examples/post-process.html)
+- Add `Colors.web` for CSS web color names to use when prototyping
 
 ## [0.0.97](https://github.com/bloopgames/toodle/releases/tag/0.0.97)